+++ gaseio/__init__.py
@@ -3,7 +3,7 @@ GASEIO
 """
 
 
-__version__ = '2.5.0'
+__version__ = '2.5.1'
 
 def version():
     return __version__